Update patch set 2
Patch Set 2:

Ah, I was thinking this is in the build project, otherwise I wouldn't have clicked the submit button :-)
Brian is right, these tools are currently used in the art project.
Personally I would prefer to not doing lint in the build process. Perhaps presubmit check is a good phase to do lint but we don't have such check yet.

Patch-set: 2