Update patch set 1
Patch Set 1:

With the encoding vector, we would have to go through the LIRs one more time. This way we're writing into 2 different places at the same time. On the other hand, we have one more write to the vector with resize() instead of reserve(), but my tests show this is still about 20% faster on the boot.oat.

The encoding vector is still used in one place (plus tests).

Patch-set: 1
Label: Verified=0