Update patch set 3
Patch Set 3:

So the real reason for the @brief that you said don't add value (constructor and destructor) are really because that way doxygen makes everything seem homogenous.

If you look at the doxygen page of art::DataflowIterator, you will see that brief will add a line under the name of the function with whatever was "briefed".

I find that if you remove the brief for a function, even if it is trivial, you get the documentation that is a bit too "compact" and not homogenous. Personal preference so, for now, I've removed it as requested.

I can provide the doxyfile I use normally, which turns on a lot of things "callee/caller graphs",references, etc. But the simplest to test would be to just go in art/compiler/dex and do "doxygen -g", then doxygen.

That will give you the base generation in a folder called html. You can then click on classes and see DataflowIterator as an example.

Let me know if you'd like my doxyfile otherwise that turns on pretty much everything I use when I'm walking around the code.

Patch-set: 3
Reviewer: Jean Christophe Beyler <1016743@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0