Update patch set 3
Patch Set 3:

Simple answer: I did not measure it. How would you go about that? For the moment, I've only run the emulator on x86/arm with ART on and made sure it booted. Do you want me to time the boot time?

Longer answer:
I agree that I've seen that kind of overhead but here, I would be surprised:

- We get a constant number of string creations when the pass driver is created
  - I didn't want to go down the singleton route to create one PassDriver across the whole execution, so we do recreate it at each CompileMethod start BUT we can then tweak the pass order per compilation unit

- We return const std::string & and catch them that way everywhere (normally)

- We only create new strings when doing the time split or the dumping
  - The latter being debug anyway
  - The former, I could change into using the char* in the const std::string with a memcpy but I'd be surprised that would be necessary

But if you give me a way to measure it, I'll be happy to give it a go.

Patch-set: 3
Label: Verified=0