Update patch set 1
Patch Set 1:

You've not ran CTS when changing how access checks are performed, you should. Your CL is about the test for final field putting, I don't see any functional difference between what's there now and what you have, although the fact we don't trust the fields_class is more explicit by moving the boolean up. To be clear, without testing this change is likely a regression. The scope of the change is to remove some ambiguity with a final field check, but it also refactors other code. There is no bug associated with this and I'd only be happy with a change like this if I knew it was dealing with all the corner cases as our current code does.

Patch-set: 1