Update patch set 2
Patch Set 2:

Lost the first part of my comment above.  Should have been something like:

"This is an example of how we fix reverted CLs.  Instead of uploading a fresh CL, we "revert the revert".  To do this, locate the "revert" or "revert change" button when you examine the CL".

Patch-set: 2