Update patch set 1
Patch Set 1: Code-Review+1

The only thing I would change potentially would be that before people could grep MIROpt, now they've lost that.

If we want to be lazy, perhaps we should just rename then the passes with MIROpt: in the names of the passes in bb_optimizations.cc/.h.

Or, perhaps we should fix that the timing supposes the pointer of the name is going to be valid across, allowing it to make a copy if need be (could be an optional boolean that we set to true in this particular case).

Just my 2 cents

Patch-set: 1
Reviewer: Jean Christophe Beyler <1016743@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1