Update patch set 3
Patch Set 3:

I consider breakpoints in getters and setters a very important use case. And we currently don't have an alternative to adding a new oat file section. So, if we want inlining at all, we need to do this or something similar.

The only alternative would be the ABI rewrite where we no longer pass the ArtMethod* to each method and have a pc->method lookup for stack traces. Are we doing this anytime soon?

Patch-set: 3