Update patch set 3
Patch Set 3:

Agreed about deoptimizing everything for a breakpoint in a getter/setter as being sub-optimal. Its better than no breakpoint. Optimizing the runtime for debug is not something I'm particularly keen on at the moment. The selective deopt's #1 use-case was that people attach the debugger and have 0 breakpoints. A full deopt then was clearly overkill and led apps to be slow.

Patch-set: 3