Update patch set 3
Patch Set 3:

I'm concerned about adding new information in the oat file which would expand its size. Could we add a special access flag on inlined method for instance? This would be a pretty quick test in the debugger (to know if we can only deoptimize the method itself or everything) and it would not impact oat file size.

Patch-set: 3