Update patch set 2
Patch Set 2:

(1 comment)

The const is just a decorator for the argument for the benefit of the implementation of the method, there's no performance implications. It has implications if we want to do things like read-barriers that will rewrite an object. For member variables (fields) const is either benign or a bug. The bug it can introduce is to not keep a loaded value in a variable but to rematerialize it by reading the const memory. Const also doesn't play nicely with reference poisoning without const casts, at which point there's so much noise in the code we're just better off without it.

Patch-set: 2