Update patch set 1
Patch Set 1:

Serguei, good catch! There is a race condition between RosAlloc::BulkFree() and RosAlloc::RevokeThreadLocalRuns() with regard to bulk_free_bit_map.

I think that the second scenario is caused by the first race condition.

I'll fill in some background info and describe my thoughts: While the patch does seem to avoid the race, there's a design implication and a performance implication here. bulk_free_bit_map is meant to be written into by the GC (BulkFree) without acquiring a lock as the comment on it in rosalloc.h describes. Given this, RevokeThreadLocalRuns(), rather than BulkFree(). is the buggy one here. Also, the loop in BulkFree() is very hot and adding the lock/unlock there decreases the allocation/GC speed. I remeasured this and an internal memory allocation performance benchmark slows down by ~40% (4009 ms -> 5590 ms). I'd like to avoid this obviously.

I revised the patch into a fix that synchronizes between BulkFree() and RevokeThreadLocalRuns() at a higher level. Since RevokeThreadLocalRuns() is a relatively rare event (thread termination, etc.) and there's a limit to the number of pointers that BulkFree() handles at a time, it seems reasonable. And importantly, with this patch, I don't see the performance regression in the same benchmark. I think this patch is simpler in terms of locking protocol and is in line with the original intention.

Do you have a test that reliably reproduces this bug? If you do, I'd appreciate it if you could run the new patch against it (who knows, I could be wrong) and even better, if you can contribute the test, since we don't have one (so that we can catch this bug now and in the future.) A guess is that the problematic RevokeThreadLocalRuns() call is coming from the thread termination code (Thread::~Thread) which could happen while GC is sweeping the heap (BulkFree()) since it looked like the other call sites are at points where threads are suspended.

Patch-set: 1