Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

Not verified but so far looks good.
A related but different issue is that support_boot_image_fixup_ in the compiler driver is false for Intel, incurring overhead if say StringBuilder calls String. We need a linker to fix this kind of code up, there's a hack for Thumb2 currently that uses the literal pool.

Patch-set: 1
Label: Code-Review=+1