Update patch set 3
Patch Set 3:

The test case is a bit difficult to create. You need a dex file defining a method but not it's declaring class. As strange as it sounds it is possible and I've seen stuff like that for some test apks. Even though we actually failed to resolve the referrer's class in that particular case, there's no reason why it can't actually be resolved -- in a different dex file.

Patch-set: 3