Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

This is a bad fix. If an app uses more than one dex file with the same name and we use the same oat location for it bad things will eventually happen.

You end up with multiple OatFiles with the same location, all referencing the same file. If dex2oat created a new oat file their file handles would actually reference different files (all but the last one actually deleted) but AFAICT we're passing a file handle to dex2oat and only overwrite the contents, so all the file handles will really refer to the same file. It is undefined what we get in the MAP_PRIVATE areas of the overwritten files. So if the app needs to read anything from not-the-last oat file, it may not get what it needs.

Patch-set: 1
Label: Code-Review=-1