Update patch set 1
Patch Set 1:

ART tries to initialize bootclasspath classes that are on the preloaded-classes list at compile time. if they call native methods, that won't work, so we have to blacklist them for now. there is a CL up in ART to avoid manual blacklist by having a way of rolling back.

Patch-set: 1