Update patch set 1
Patch Set 1:

Please have a look, and let me know what you think.  This is quite rough at this point, and suggestions to improve would help.

I would like to know:
- How to do this for debug builds only
- Suggestions for improvements

In order to get better information about functions, such as types, parameters, line numbers, etc, much more work would have to be done.  I am not sure that doing this in oat_writer and elf_writer_quick is the right place.

Some of this is machine independent, and some machine dependent.  In particular, the updating of the debug information before passing it to GDB has knowledge of the format of the debug_info.  Any better way to handle this?  Start parsing debug_abbrev and debug_info looking for low_pc and high_pc to update?

Patch-set: 1
Label: Verified=0