Update patch set 1
Patch Set 1:

This should probably use a command line switch --gdb to interface with GDB when we want to do so.  That would let us leave the code in all the time.  It would depend on how much more space is being used, and how much extra time it takes to generate the extra information on the OAT file

Patch-set: 1