Update patch set 4
Patch Set 4:

| Bill, I haven't found any local value numbering in dalvik/vm/ . Is that actually done on dalvik?
No, but there is a dex code pattern match in the JIT which I suspect did not consider the possibility of aliasing.  It would be a bit tricky to construct a test to trigger it (you'd have to execute the code path enough times without throwing an exception to trigger a trace translation first, and then inject the NULL reference).   I'll file a bug against myself to find out whether a bug really exists there.

Patch-set: 4
Label: Verified=0