Update patch set 9
Patch Set 9: Code-Review-1

The issue is that the notion of what kind of fast/slow path is used is an agreement between the verifier, the compiler "interface" and the runtime support. This change moves an amount of that into mir_annotations meaning there are more moving parts to have correct, in mir annotations, for example were we to add a JIT or trace a back a bug. The compiler, quick in this case, should be querying the compiler "interface" (driver) for what behavior it is allowed to get away with.

Patch-set: 9
Reviewer: Ian Rogers <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1