Update patch set 2
Patch Set 2:

So it's Bill's call. For me the ir_structures.cc doesn't make sense and isn't following ART's general pattern (but for that matter neither is most of the quick compiler). My reason for saying I don't follow the sense is that ir_structures.cc is defining member functions for MIR, basic block and child block iterator. The other definitions for these things are in other files, and so going from declaration to definition is going require the use of grep.

Patch-set: 2