Update patch set 2
Patch Set 2:

Well this was a first patch to move these here and then potentially move the others as we go. I did not want to create a huge patch that might conflict with a lot of patches (especially if the decision was no to the patch). 

My rational is that as we add to these structures, they are going to get more API utilitarian functions and I think they belong more in a infrastructure related code file than bloating mir_graph.cc where a lot of functions have gone.

There is already 1000+ lines for mir_graph.cc; I was trying to keep it at that and put anything really utilitarian (removing a MIR, add before/after, what BB is a MIR from, rewriting a MIR, replace a child BB with another, ...) there and not in a generic catch all mir_graph file.

But if you don't agree with this, just let me know where you would put these functions.

Ps: I also use ctags generally so I don't have that issue of greping all over the place :)

Patch-set: 2
Reviewer: Jean Christophe Beyler <1016743@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0