Update patch set 2
Patch Set 2:

(6 comments)

no, not a part of a different CL, part of chasing making RuntimeOptions const in dex2oat. It was non-const to push with compilercallbacks. When I saw that that impl of CompilerCallbacks was cut/paste between dex2oat and common_test.h I pulled it out.

Patch-set: 2