Update patch set 5
Patch Set 5:

I haven't been in the discussion, and this is a WIP. So, I may be missing something, but IIUC, since StringFactory uses JNI to allocate string objects, could this change slow down string allocations (even if this new string may save an extra internal char array allocation)? I'd worry as string allocations are common (and used a lot in Ritzperf.) The StringFactory native calls may need to be turned into compiler entry points to maintain (if not improve) the allocation performance. Is that in the plans?

Patch-set: 5
Label: Verified=0