Update patch set 1
Patch Set 1:

Confirming the portable build does not compile for me either now... That's surprising as I was pretty sure I was able to compile it with this CL applied. Running it was the only (expected) problem. I'll fix the build on Monday.

Patch-set: 1
Reviewer: Nicolas Geoffray <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0