Update patch set 1
Patch Set 1:

(4 comments)

Thanks Brian. PTAL. I wasn't sure if host-prefix wasn't used in weird ways throughout the build system, but it looks like we can remove it.

I also removed kOatLocation, but that means the image file has now a different ABI than before. Is that ok because it's just the image file, or should we bump a version somewhere?

Patch-set: 1