Update patch set 1
Patch Set 1:

Unless you are testing on a mac this code change has no effect. From mutex.h:

#if defined(__APPLE__)
#define ART_USE_FUTEXES 0
#else
#define ART_USE_FUTEXES 1
#endif

The Linux code was around for the benefit of MIPS that had a bug in its syscall implementation that was causing problems in the ART_USE_FUTEXES code. If you want to test this properly I'd recommend modifying mutex.h.

Patch-set: 1
Label: Verified=0