Update patch set 1
Patch Set 1:

GCC regressed its warnings. There was an argument for a better GCC for the target but for the host we wanted to have better warnings. For the OS/X build we've switched the host to using clang. I believe GCC will be bringing back -Wthread-safety at which point I'll remove my objections. For host OS/X building GCC isn't an option.

Patch-set: 1