Update patch set 1
Patch Set 1:

Ben, that misses the point. We upgraded the target GCCs and lost thread-safety support. We warn about this. This change regresses the host GCC and doesn't even warn about it. It was a deliberate decision not to do for the host what we do for the target. We shouldn't be regressing our only build configuration that detects thread safety violations as these happen frequently in our code base.
A viable alternative would be to do this change, but add host warnings about the regression in GCC and if our host clang supports the thread-safety annotations, switch to building our host builds to that.  We tend to run debug versions of ART on the host as host performance isn't critical. All the more reason we don't need to upgrade it if its a regression in other areas, such as generated warnings.

Patch-set: 1