Update patch set 1
Patch Set 1:

Is the issue here that we don't pull in the global defines needed by art/Android.mk because the dex_preopt.mk is only pulled in when you build a java project? seems like we are defining it in the wrong place then.

since the DEX2OAT_TARGET_INSTRUCTION_SET_FEATURES is just a derivative of TARGET_CPU_VARIANT, it might seem best to move it to where those are defined, but looking that is too widespread.

can we move to just having the the definition be in an include that is included by dex_preopt_libart.mk as well as art/ with multiple include guard?

Patch-set: 1