Update patch set 1
Patch Set 1:

Bill,

Thanks for the quick reply, for the feedback and for sharing your plans. It is reassuring to hear that the patch is roughly aligned with what you have intended to do. No worries about merging. This is very much an early experiment. My plan is to continue to go through the missing instructions in the encoding map and to complete it. In the meanwhile, I am keeping an eye on what you are doing upstream. So far rebasing has been straightforward, which suggests that we have been moving along compatible directions. Also, our aim is to help making Art run on Aarch64 as soon as possible. As long as this happens, we are not too concerned if parts of our work do not get merged. Obviously, it is better if we end up helping you as much as we can.

Patch-set: 1
Label: Verified=0