Update patch set 1
Patch Set 1: Code-Review-2

This isn't the first request for this.

This is a prebuilt clang toolchain bug that we are looking. The problem is that it is using /usr/bin/ld and not a prebuilt. It works on ubuntu 12.04 if /usr/bin/ld is ld.bfd, but not of it is ld.gold.

Patch-set: 1
Reviewer: Brian Carlstrom <1003723@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2