Update patch set 1
Patch Set 1: Code-Review-1

The trouble is that Memmove and Memcpy is logically considered as backward and forward copying functions. However when we need forward copying but arrays are still overlapped we cannot use memcpy function due to its specification. But I'm ok if memcpy is more efficient I will add the special check whether arrays are overlapped and use memcpy or memmove correspondingly.

Patch-set: 1
Label: Code-Review=-1