Update patch set 2
Patch Set 2:

The condition is dst_pos < src_pos but src_pos - dst_pos < count. In this case arrays are overlapped but we need forward copying. I guess I will have a trouble to provide a test which fails because to reproduce the issue I need memcpy function worked incorrectly with overlapped arrays. An example of this implementation can be bionic/libc/arch-x86/string/ssse3-memcpy-atom.S which compatible with specification of memcpy and does not work correctly with overlapped array. But x86 emulator does not use sse3.

Patch-set: 2