Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

This change mixes the notions of registers between quick and those in utils. The utils assemblers are consistently used elsewhere in using the runtime and managed register notions of registers. I would like to avoid aliasing the notion of register between quick and the utils assembler.

Patch-set: 3
Label: Code-Review=-1