Update patch set 1
Patch Set 1: Code-Review-1

Please do not merge the code. There is bug with the code. Sorry for the stupid mistake I've made.

It doesn't handle the case when dest and src overlaps. It affects both div remainder and multiply. I will push the fixed change-set, as soon as I can.

Patch-set: 1