Update patch set 2
Patch Set 2:

Your point is about when we generate oat files, this patch is about when we load an oat file allowing it to be executable when the architecture types don't match.
For your change you need to modify ClassLinker::FindOrCreateOatFileForDexLocation to not regenerate in the case where not matching ISA is an error. ie this code is correct for the location its in.
I don't agree that we shouldn't regenerate in that case for robustness reasons.
I think you are trying to diagnose bugs in the package manager by using this bug as a feature, you can get the same information by examining logcat output so I fail to see why we need the bug.

Patch-set: 2