Update patch set 1
Patch Set 1:

Although it makes the title unwieldy, we generally try to upload fixes to reverted CLs as reverts of the revert.  That way, it preserves the history and makes it easier to tell what changed since the revert.  The way you'd do that is create a revert of the revert, cherry-pick that into your tree and then add your fixes.

Patch-set: 1