Update patch set 1
Patch Set 1:

We have to.  The selinux rules are such that apps cannot write into /data/dalvik-dev.  That is written by installd and those processes run by it.  Also, installd cannot write into /data/data which is where the apps want to write.  This was thrashed out with nnk (the owner of selinux for Android), and the outcome is that installd creates the file with group write for the app and the app reads and writes it.

We went round and round with this for weeks and this is a solution that partially satisfies everyone.  SELinux is happy because it doesn't open up /data/data for installd to write to.  Apps are happy because they have somewhere to write their profile to.

Patch-set: 1