Update patch set 3
Patch Set 3:

(2 comments)

This is indeed about Bug: 11767815 as I'm trying to move the mapping table offset and vmap table offset to OatMethodHeader, https://android-review.googlesource.com/#/c/90699/ . I'd like to have the offsets there unsigned and separating the maps is the way to do it.

This doesn't reduce the size of the maps (Bug: 9437697) per se, it just reduces the size of the code alignment padding. This is a nice side-effect based on simple statistics. When the padding is more or less randomly distributed across integers in range [0, 16) we can expect the average padding to be about 7.5 bytes. However, when we move the maps to their own section, the padding is always even (Thumb) with a bias towards a multiple of 4 (whenever we have a non-empty literal pool) and that pushes the average down (for even paddings about 7 bytes, for multiples of 4 about 6 bytes, so it should be somewhere in between 6 and 7).

Patch-set: 3
Label: Verified=0