Update patch set 1
Patch Set 1:

Hi Nicolas,

This bug is related to access to the memory area, which we have no rights to access to. So, it might lead to crash or not, print incorrect result or not, depends on many factors. I have an reproducer, but I'm not sure it'll reproduce this bug on all configurations.

Patch-set: 1