Update patch set 1
Patch Set 1: Code-Review+1

In order to get the Loop formation system up, we added a lot of infrastructure in the ME. This patch is one of them that adds to the ME what is required for the MIR, the SSARepresentation and just what is needed for the BB from the MIR implementation.

Use-def chains are really used a lot throughout the loop infrastructure and creation, especially the IV handling.

Patch-set: 1
Label: Code-Review=+1