Update patch set 1
Patch Set 1:

Here is an initial version of a system enabling a .so to be loaded and change the behavior of the passes. It actually is useful to work on passes separately without recompiling the whole system but also debugging (turning off a pass via a .so) and instrumentation (adding a sanity check pass via a .so).

There is probably a security question and then a compatibility question to be handled when hand-shaking the loading up of the .so but we can turn off the feature by simply guarding the code in parsed_options.cc (via an option/setting) for example and still have the functionality available when necessary.

Patch-set: 1
Reviewer: Jean Christophe Beyler <1016743@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0