Update patch set 1
Patch Set 1:

Hmm, I wanted this to always appear when we get a signal that we don't catch.  That way we can see that our handler was invoked and it's not a mystery where the signal came from.

Is there a way to use LOG without causing an allocation?  Perhaps __android_log_print?

Patch-set: 1
Label: Code-Review=0