Update patch set 2
Patch Set 2:

But if isn't the C library's handler that it called we don't see anything.  I don't want to lose the information that we trapped a SEGV and passed it on because we didn't need it.  It's not that important but it's been useful in the past couple of instances - like we saw that facebook didn't cause our handler to be invoked - useful information.  I'd not like to have to guess if our handler was called or not.

Patch-set: 2
Reviewer: Dave Allison <1036486@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0