Update patch set 3
Patch Set 3:

adding bug number for context.

i think the problem in that bug is that because of some other issue (stack overflow?) we ended up getting a SIGSEGV do to this logging call in the signal handler, which prevented the libc handler from being called. I think its a lot more appropriate to turn logging on when debugging an issue than to leave it on by default. in the case of the signal handlers, it seems better to be even more cautious about this?

Patch-set: 3
Label: Verified=0