Update patch set 1
Patch Set 1:

(1 comment)

The block entry ssa map was the main thing that caught my eye on this.  I am not absolutely opposed to increasing the memory footprint, but any CL with a significant memory increase should be accompanied by an argument of what (i.e. performance, capability) the memory expenditure will buy us.  In this case, rebuilding the entry map on an as-needed basis sounds right.  If, in the future, the rebuilding cost starts showing up in profiles we can revisit.

Patch-set: 1
Label: Code-Review=0