Update patch set 2
Patch Set 2:

The changes to the Quick compiler to get us back to functional state are minor.  However, part of the motivation for making this change at this time was the realization that the Quick compiler is not currently handling reference width consistently.  On 64-bit targets, the Dalvik virtual register reference width is 32, while the physical register reference width is 64.  Handling this with more precision won't be a huge change (just some relatively minor tweaking of the register bookkeeping), but would get a bit messy if we had to treat Method* as a special case (i.e. 64-bits both in-register and in-frame).  By shifting Method* back to 32 bits when stored in frame, we can treat it just as any other reference.

Note that in the future (with the optimizing compiler) we don't expect to have a restriction that references are always stored compressed - it should be free to spill uncompressed references using natural width.   We expect a revision of the runtime model when it comes online, and can revisit this issue then.

Patch-set: 2
Reviewer: Bill Buzbee <1001578@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0