Update patch set 4
Patch Set 4:

I am interested in some input. The assembler method used to generate the call to the native code was created with the intention of calling *(*(sp + offset1) + offset2), where the JNI compiler feeds offset1 so that the ArtMethod is read from the stack and offset2 is the offset to the jni entrypoint.

With 64b, this is not true anymore, and I had to change the code in both 64b assemblers. Now, while the method is still named Call, it will do a 32b load followed by a 64b load, which is not obvious from the signature.

On all architectures the call is actually done in at least two steps, handling the two loads. On x86, the second load is coalesced with the call, but on all archs the first load is separate. I wonder whether it would be better from a code maintenance and readability point to take the Call apart in the jni compiler - make it an (explicit 32b) load and an offset call.

Patch-set: 4