Update patch set 4
Patch Set 4: Code-Review+1

The interpreter seems to work when running on the model in Arm64 (I only ran a few tests on the AEM model). The quick compiler needs to be adapted accordingly, but this is not what this patch is concerned with.

Patch-set: 4
Reviewer: Matteo Franchin <1037505@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1