Update patch set 2
Patch Set 2:

For Intel instructions the opcodes are either "word" or "byte". In our case word is always 32-bit or 4 bytes. The CHECK is that our immediate is 4-bytes, which is sound as we filtered the 1 byte case immediate before this. The operand size override prefix is then stating that the "word" sized thing should be modified. The modification is always from 4-bytes to 2-bytes. 8-byte immediates are only possible with a special opcode, and are not general purpose (nor do they use prefixes). If this logic were broken it would also be broken for register dumping, see DumpReg0, specifically:
  size_t size = byte_operand ? 1 : (size_override == 0x66 ? 2 : (rex_w ? 8 :4));

Patch-set: 2