Update patch set 1
Patch Set 1:

(2 comments)

> We should just tell the debugger to "get off of my heap.." Really, its a debugger so we care little about performance and these things can live in JNI.

Good point. That would make the need to instrument the debugger weak roots go away as JNI (weak) globals are already instrumented.

I think a key issue is that ObjectRegistry uses a std::map whose key is mirror::Object* to keep track of mirror::Object* to ObjectId mappings.

It's not clear if we can avoid using a map whose key includes some sort of the identity of the object.

We could consider using jobject as the key type with JobjectComparator. But, it might be broken in that if we use the JNI weak global as the key type, which is likely what we want as we don't make them strong refs, the comparator isn't going to be stable in that its behavior changes when a weak global is cleared, which, I think, would break the map.

If we use the identity hash code as the key type, the same issue may apply.

Also, we might overflow the global indirect reference table if we track a lot of objects from the debugger.

Not knowing much about the debugger, I'm not sure if there's a good solution.

Sebastien, can you help?

Patch-set: 1