Update patch set 1
Patch Set 1:

I'm concerned we'd need to update the whole object registry table each time we add an element. I think most of the "adds" are done in the beginning of the debugging session, when the debugger looks for classes and particular objects so that would slow down the warm up time. Besides, we'll need to also add fields and methods to this registry wrt moving collector so the table would contain more elements than it does now.

By default, Object* is added to the object registry with a JDWP id and a JNI weak global ref. Therefore it can be garbage collected even if the debugger knows its JDWP id. To be sure an object is not collected, the debugger can ask us to disable collection for this object (see JDWP ObjectReference.DisableCollection command, http://docs.oracle.com/javase/7/docs/platform/jpda/jdwp/jdwp-protocol.html#JDWP_ObjectReference_DisableCollection) in which case we promote the JNI weak global ref to a JNI global ref (we create a global ref and delete the weak ref). The debugger can also enable the collection in which case we demote the global ref to a weak ref. I think an Object* being garbage collected should be removed from the ObjectRegistry maps. I'm not sure we're doing this right now, only when the debugger explicitly asks us to do so with Jdwp VirtualMachine.DisposeObjects command.

The essential point here is every Object* must have a unique JDWP identifier which remains the same during the debugging session. We need to keep a consistent mapping between Object* and JDWP ids. That is to say we must be able to find the Object* associated with a JDWP id *and* find the JDWP id associated with an Object*. If we change the map key, we need to use something that does not change.

Patch-set: 1