Update patch set 1
Patch Set 1: Code-Review-1

This doesn't look like the right fix. You are changing the instruction depending on whether the register is a byte register and the build is 64-bit, however, the changed instruction is the same instruction. The issue is the use of a byte register. The opcode and the register already inform you as to whether this will need a REX prefix, can't we just make byte registers do the right thing? This will have the effect of not introducing 2x the byte instructions and fixing the byte instructions this change neglects to fix.

Patch-set: 1
Reviewer: Ian Rogers <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1